Divided ClockFace.jsx into smaller files for ease of editing
